+++ test/chart_engine_series_test.dart
@@ -1,3 +1,5 @@
+@TestOn('browser')
+
 import 'package:chart_engine/src/chart_engine_series.dart';
 import 'package:test/test.dart';